+++ packages/plugins/contracts/test/conflictingNameContracts/First.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.0;
 
 contract ConflictingNameContract {
-    uint public number;
+    uint256 public number;
 
-    constructor(uint _number) {
+    constructor(uint256 _number) {
         number = _number;
     }
 
-    function set(uint _number) public {
+    function set(uint256 _number) public {
         number = _number;
     }
 }
